==> Proyecto/Scripts/Notebooks/5 DatabaseCreation.ipynb <==
## Trabajo Fin de Máster <br/> Diseño de una arquitectura multimodal para descripción textual de pares imagen-audio

## Script 5. Creación de la base de datos completa

En este notebook, unimos las bases de datos de imagen y audio, creando un dataframe (posteriormente guardado en un csv) en el que emparejamos imágenes con audios, sin seguir ningún tipo de criterio.

Dado que hay muchos menos audios que imágenes, lo que se hace es asignar el mismo audio a varias imágenes, cosa que hacemos recorriendo la lista de audios en ciclo.

### Paso 1. Lectura de las bases de datos de entrenamiento

Leemos los documentos csv en los que guardamos la información sobre las bases de datos de imagen y audio.

import os
os.chdir("..")
os.getcwd()

import pandas as pd

images = pd.read_csv('Final_Database/image/imagesDB_train.csv')
audios = pd.read_csv('Final_Database/audio/audioDB_train.csv')

images.head()

audios.head()

### Paso 2. Fusión de los dataframes

final_db = images.copy()

from itertools import cycle

for clase in audios.classID.unique():
    selected_rows = final_db.classID==clase
    audios_ciclo = cycle(audios[audios.classID==clase].audio_path.unique())
    final_db.loc[selected_rows, 'audio_path'] = [next(audios_ciclo) for _ in range(selected_rows.sum())]

final_db

final_db.to_csv('Final_Database/finalDB_train.csv', index=False)

### Paso 3. Repetimos con los datos de test

images = pd.read_csv('Final_Database/image/imagesDB_test.csv')
audios = pd.read_csv('Final_Database/audio/audioDB_test.csv')

final_db = images.copy()

for clase in audios.classID.unique():
    selected_rows = final_db.classID==clase
    audios_ciclo = cycle(audios[audios.classID==clase].audio_path.unique())
    final_db.loc[selected_rows, 'audio_path'] = [next(audios_ciclo) for _ in range(selected_rows.sum())]

final_db

final_db.to_csv('Final_Database/finalDB_test.csv', index=False)
